% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{calculateRFM}
\alias{calculateRFM}
\title{caculateRFM}
\usage{
calculateRFM(transactions)
}
\arguments{
\item{transactions}{data table which contains the transactions}
}
\value{
RFM table
}
\description{
Calculates recency, frequency and monetary values for every customer.
}
\details{
\code{transactions} has to contain the columns \code{PurchAmount}, \code{TransDate}.
}
