\name{fm-package}
\alias{fm-package}
\alias{fm}
\docType{package}
\title{
\packageTitle{fm}
}
\description{
\packageDescription{fm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fm}
\packageIndices{fm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{fm}

Maintainer: \packageMaintainer{fm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
