% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRank.R
\name{getRank}
\alias{getRank}
\title{getRank}
\usage{
getRank(x, depth, f)
}
\arguments{
\item{x}{Either Recency, Frequency, or Monetary.}

\item{depth}{an integer. The rank (i.e. given the rank, the corresponding cluster is returned).}

\item{f}{a function. Either the maximum or minimum.}
}
\description{
This function computes finds the ranking for a single cluster.
}
\examples{
getRank()
}
\keyword{RFM,}
\keyword{Rank,}
\keyword{kmeans++}

