% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRankedClusters.R
\name{getRankedClusters}
\alias{getRankedClusters}
\title{getRankedClusters}
\usage{
getRankedClusters(data, listIDs, colname1, colname2, f)
}
\arguments{
\item{data}{a dataset with columns named Date, Recency, Frequency, and Monetary.}

\item{listIDs}{a list of vectors of User IDs corresponding to their respective clusters.}

\item{colname1}{.}

\item{colname2}{.}

\item{f}{a function. Either the maximum or minimum is used (depending on which variable is being ranked).}
}
\description{
This function ranks the results of clusterings.
}
\examples{
getRankedClusters()
}
\keyword{RFM,}
\keyword{Ranking,}
\keyword{kmeans++}

