\name{rfm}
\alias{rfm}
\title{This package create a rfm model for customers segmentation}

\description{
This package use an input of data to create a RFM segmentation (where R = recency, F = frequency and M= monetary).
It will identify groups of customers in marketing database in order to target marketing activity
}
