% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDates.R
\name{randomDates}
\alias{randomDates}
\title{randomDates}
\usage{
randomDates(N_Dates, startDate, endDate)
}
\arguments{
\item{N_Dates}{integer Number of random dates to draw.}

\item{startDate}{character Beginning of date range (for example "2015/01/01").}

\item{endDate}{character End of date range (for example "2016/12/31").}
}
\description{
This function generates random dates.
}
\examples{
randomDates(10,"2016/01/01","2016/12/31")
randomDates()
}
\keyword{Dates}
\keyword{Random}

