% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm_avg_segment.R
\name{rfm_avg_segment}
\alias{rfm_avg_segment}
\title{rfm_avg_segment fucntion}
\usage{
rfm_avg_segment(model)
}
\arguments{
\item{model}{a data frame loaded with rfm_load or created previously}
}
\value{
a Data.frame object
}
\description{
Summarise the average value for each existing score
}

