% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-data.R
\docType{data}
\name{rfm_data}
\alias{rfm_data}
\title{RFM Data Set}
\format{A tibble with 25613 rows and 4 variables:
\describe{
  \item{order_date}{order date}
  \item{order_id}{order id}
  \item{customer_id}{customer id}
  \item{revenue}{transaction amount}
}}
\source{
https://github.com/joaolcorreia/RFM-analysis
}
\usage{
rfm_data
}
\description{
A dataset containing transactions of different customers.
}
\keyword{datasets}
