% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-data.R
\docType{data}
\name{rfm_data_orders}
\alias{rfm_data_orders}
\title{RFM transaction data}
\format{
A tibble with 49.6 rows and 3 variables:
\describe{
  \item{order_date}{Order date.}
  \item{customer_id}{Customer id}
  \item{revenue}{Transaction amount.}
  \item{first_name}{First name of the customer.}
  \item{last_name}{Last name of the customer.}
  \item{email}{email id of the customer.}
}
}
\usage{
rfm_data_orders
}
\description{
A dataset containing transactions of different customers.
}
\keyword{datasets}
