% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFM.R
\name{rfm_function}
\alias{rfm_function}
\title{get_max_PurchAmount}
\usage{
rfm_function(
  table,
  time_format = "\%d.\%m.\%Y",
  weight_frequency = 1/3,
  weight_recency = 1/3,
  weight_monetary_value = 1/3
)
}
\arguments{
\item{table}{a data object with columns TransDate, Customer, PurchAmount, Cost}

\item{date}{a specification for the time of interest. Must be of format "dd.mm.yyy"}
}
\value{
table with RFM scores and corresponding group for each Customer
}
\description{
Function that return weighted rfm scores for customers
}
\details{
\code{table} contains the transaction data. The data set must must be a data.table and contain a
            column labeled "Customer" that allows unique customer identification
            and a column labeled "TransDate", indicating the purchase date.
            The column "PurchAmount" specifies the total spending per purchase.
}
\examples{
rfm_function(table, date="01.10.2020", weight_frequency = 0.2,
weight_recency = 0.2, weight_monetary_value = 0.6)

}
