% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_heatmap}
\alias{rfm_heatmap}
\title{RFM Heatmap}
\usage{
rfm_heatmap(data)
}
\arguments{
\item{data}{an object of class \code{rfm_table}}
}
\value{
Heatmap
}
\description{
The heat map shows the average monetary value for different 
categories of recency and frequency scores. Higher scores of frequency and 
recency are characterized by higher average monetary value as indicated by 
the darker areas in the heatmap.
}
\examples{
# rfm table
analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# heat map
rfm_heatmap(rfm_result)

}
