% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots-data.R
\name{rfm_hist_data}
\alias{rfm_hist_data}
\title{Histogram data}
\usage{
rfm_hist_data(rfm_table)
}
\arguments{
\item{rfm_table}{An object of class \code{rfm_table}.}
}
\description{
Data for generating histograms.
}
\examples{
# using transaction data
analysis_date <- as.Date('2006-12-31')
rfm_order <- rfm_table_order(rfm_data_orders, customer_id, order_date,
revenue, analysis_date)

# histogram data
rfm_hist_data(rfm_order)

# using customer data
analysis_date <- as.Date('2007-01-01')
rfm_customer <- rfm_table_customer(rfm_data_customer, customer_id,
number_of_orders, recency_days, revenue, analysis_date)

# histogram data
rfm_hist_data(rfm_customer)

}
