% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-launch-shiny-app.R
\name{rfm_launch_shiny_app}
\alias{rfm_launch_shiny_app}
\title{Launch shiny app}
\usage{
rfm_launch_shiny_app()
}
\description{
Launches shiny app.
}
\examples{
\dontrun{
rfm_launch_shiny_app()
}
}
