% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm_load.R
\name{rfm_load}
\alias{rfm_load}
\title{rfm_load fucntion}
\usage{
rfm_load(file)
}
\arguments{
\item{file}{Path of the file that need to}
}
\value{
a Data.frame object
}
\description{
Load data in to create model
}

