% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm_model.R
\name{rfm_model}
\alias{rfm_model}
\title{rfm_load fucntion}
\usage{
rfm_model(df)
}
\arguments{
\item{df}{a data frame loaded with rfm_load or created previously}
}
\value{
a Data.frame object
}
\description{
Load data in to create model
}

