% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm_plot.R
\name{rfm_plot}
\alias{rfm_plot}
\title{rfm_plot fucntion}
\usage{
rfm_plot(model)
}
\arguments{
\item{model}{the results of rfm function -- a rfm model data.frame}
}
\value{
a serie of graphs
}
\description{
Plot the basic info after model is done
}

