% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-segment.R
\name{rfm_plot_median_recency}
\alias{rfm_plot_median_recency}
\alias{rfm_plot_median_frequency}
\alias{rfm_plot_median_monetary}
\title{Segmentation plots}
\usage{
rfm_plot_median_recency(
  rfm_segment_table,
  color = "blue",
  sort = FALSE,
  ascending = FALSE,
  flip = FALSE,
  plot_title = NULL,
  xaxis_label = NULL,
  yaxis_label = NULL,
  font_size = 6,
  interactive = FALSE,
  print_plot = TRUE
)

rfm_plot_median_frequency(
  rfm_segment_table,
  color = "blue",
  sort = FALSE,
  ascending = FALSE,
  flip = FALSE,
  plot_title = NULL,
  xaxis_label = NULL,
  yaxis_label = NULL,
  font_size = 6,
  interactive = FALSE,
  print_plot = TRUE
)

rfm_plot_median_monetary(
  rfm_segment_table,
  color = "blue",
  sort = FALSE,
  ascending = FALSE,
  flip = FALSE,
  plot_title = NULL,
  xaxis_label = NULL,
  yaxis_label = NULL,
  font_size = 6,
  interactive = FALSE,
  print_plot = TRUE
)
}
\arguments{
\item{rfm_segment_table}{Output from \code{rfm_segment}.}

\item{color}{Color of the bars.}

\item{sort}{logical; if \code{TRUE}, sort metrics.}

\item{ascending}{logical; if \code{TRUE}, sort metrics in ascending order.}

\item{flip}{logical; if \code{TRUE}, creates horizontal bar plot.}

\item{plot_title}{Title of the plot.}

\item{xaxis_label}{X axis label.}

\item{yaxis_label}{Y axis label.}

\item{font_size}{Font size for X axis text.}

\item{interactive}{If \code{TRUE}, uses \code{plotly} as the visualization
engine. If \code{FALSE}, uses \code{ggplot2}.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Segment wise median recency, frequency & monetary value plot.
}
\examples{
# analysis date
analysis_date <- as.Date('2006-12-31')

# generate rfm score
rfm_result <- rfm_table_order(rfm_data_orders, customer_id, order_date,
revenue, analysis_date)

# segment names
segment_names <- c("Champions", "Potential Loyalist", "Loyal Customers",
                   "Promising", "New Customers", "Can't Lose Them",
                   "At Risk", "Need Attention", "About To Sleep", "Lost")

# segment intervals
recency_lower <-   c(5, 3, 2, 3, 4, 1, 1, 1, 2, 1)
recency_upper <-   c(5, 5, 4, 4, 5, 2, 2, 3, 3, 1)
frequency_lower <- c(5, 3, 2, 1, 1, 3, 2, 3, 1, 1)
frequency_upper <- c(5, 5, 4, 3, 3, 4, 5, 5, 3, 5)
monetary_lower <-  c(5, 2, 2, 3, 1, 4, 4, 3, 1, 1)
monetary_upper <-  c(5, 5, 4, 5, 5, 5, 5, 5, 4, 5)

# generate segments
segments <- rfm_segment(rfm_result, segment_names, recency_lower,
recency_upper, frequency_lower, frequency_upper, monetary_lower,
monetary_upper)

# plots
# visualize median recency
rfm_plot_median_recency(segments)

# plotly
rfm_plot_median_recency(segments, interactive = TRUE)

# sort in ascending order
rfm_plot_median_recency(segments, sort = TRUE, ascending = TRUE)

# default sorting is in descending order
rfm_plot_median_recency(segments, sort = TRUE)

# horizontal bars
rfm_plot_median_recency(segments, flip = TRUE)

# median frequency
rfm_plot_median_frequency(segments)

# median monetary value
rfm_plot_median_monetary(segments)

}
