% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-segment.R
\name{rfm_plot_revenue_dist}
\alias{rfm_plot_revenue_dist}
\title{Revenue distribution}
\usage{
rfm_plot_revenue_dist(x, flip = FALSE, print_plot = TRUE)
}
\arguments{
\item{x}{An object of class \code{rfm_segment_summary}.}

\item{flip}{logical; if \code{TRUE}, creates horizontal bar plot.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.

analysis_date <- as.Date('2006-12-31')
rfm_result <- rfm_table_order(rfm_data_orders, customer_id, order_date,
revenue, analysis_date)

segment_names <- c("Champions", "Loyal Customers", "Potential Loyalist",
  "New Customers", "Promising", "Need Attention", "About To Sleep",
  "At Risk", "Can't Lose Them", "Lost")

recency_lower <- c(4, 2, 3, 4, 3, 2, 2, 1, 1, 1)
recency_upper <- c(5, 5, 5, 5, 4, 3, 3, 2, 1, 2)
frequency_lower <- c(4, 3, 1, 1, 1, 2, 1, 2, 4, 1)
frequency_upper <- c(5, 5, 3, 1, 1, 3, 2, 5, 5, 2)
monetary_lower <- c(4, 3, 1, 1, 1, 2, 1, 2, 4, 1)
monetary_upper <- c(5, 5, 3, 1, 1, 3, 2, 5, 5, 2)

segments <- rfm_segment(rfm_result, segment_names, recency_lower,
recency_upper, frequency_lower, frequency_upper, monetary_lower,
monetary_upper)

segment_overview <- rfm_segment_summary(segments)
rfm_plot_revenue_dist(segment_overview)}
}
\description{
Customer and revenue distribution by segments.
}
