% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-segment.R
\name{rfm_plot_segment_scatter}
\alias{rfm_plot_segment_scatter}
\title{Segment Scatter Plots}
\usage{
rfm_plot_segment_scatter(
  segments,
  x = "monetary",
  y = "recency",
  xaxis_label = NULL,
  yaxis_label = NULL,
  plot_title = NULL,
  legend_title = NULL,
  print_plot = TRUE
)
}
\arguments{
\item{segments}{Output from \code{rfm_segment}.}

\item{x}{Metric to be represented on X axis.}

\item{y}{Metric to be represented on Y axis.}

\item{xaxis_label}{X axis label.}

\item{yaxis_label}{Y axis label.}

\item{plot_title}{Title of the plot.}

\item{legend_title}{Title of the plot legend.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\value{
Scatter plot.
}
\description{
Generate scatter plots to examine the relationship between
recency, frequency and monetary value.
}
\examples{
# analysis date
analysis_date <- as.Date('2006-12-31')

# generate rfm score
rfm_result <- rfm_table_order(rfm_data_orders, customer_id, order_date,
revenue, analysis_date)

# segment names
segment_names <- c("Champions", "Loyal Customers", "Potential Loyalist",
  "New Customers", "Promising", "Need Attention", "About To Sleep",
  "At Risk", "Can't Lose Them", "Lost")

# segment intervals
recency_lower <- c(4, 2, 3, 4, 3, 2, 2, 1, 1, 1)
recency_upper <- c(5, 5, 5, 5, 4, 3, 3, 2, 1, 2)
frequency_lower <- c(4, 3, 1, 1, 1, 2, 1, 2, 4, 1)
frequency_upper <- c(5, 5, 3, 1, 1, 3, 2, 5, 5, 2)
monetary_lower <- c(4, 3, 1, 1, 1, 2, 1, 2, 4, 1)
monetary_upper <- c(5, 5, 3, 1, 1, 3, 2, 5, 5, 2)

# generate segments
segments <- rfm_segment(rfm_result, segment_names, recency_lower,
recency_upper, frequency_lower, frequency_upper, monetary_lower,
monetary_upper)

# generate plots
rfm_plot_segment_scatter(segments, "monetary", "recency")
rfm_plot_segment_scatter(segments, "monetary", "frequency")
rfm_plot_segment_scatter(segments, "frequency", "recency")

}
