% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-segment.R
\name{rfm_plot_segment_summary}
\alias{rfm_plot_segment_summary}
\title{Visulaize segment summary}
\usage{
rfm_plot_segment_summary(x, metric = NULL, sort = FALSE, print_plot = TRUE)
}
\arguments{
\item{x}{An object of class \code{rfm_segment_summary}.}

\item{metric}{Metrics to be visualized.}

\item{sort}{Sort metrics in ascending or descending order.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\description{
Generates plots for customers, orders, revenue and average order value for each segment.
}
