% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFM_Segments.R
\name{RFM_Segments}
\alias{RFM_Segments}
\title{RFM_Segments}
\usage{
RFM_Segments(data, N_Recency_Clusters, N_Frequency_Clusters,
  N_Monetary_Clusters)
}
\arguments{
\item{data}{a dataset with columns named Date, Recency, Frequency, and Monetary.}

\item{N_Recency_Clusters}{an integer indicating the number of Recency Clusters to be computed.}

\item{N_Frequency_Clusters}{an integer indicating the number of Frequency Clusters to be computed.}

\item{N_Monetary_Clusters}{an integer indicating the number of Monetary Clusters to be computed.}
}
\description{
This function computes the RFM_Segmentation for a dataset with columns named Date, Recency, Frequency, and Monetary.
}
\examples{
RFM_Segments()
}
\keyword{RFM,}
\keyword{kmeans++}

