% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-table.R
\name{rfm_table}
\alias{rfm_table}
\alias{print.rfm_table}
\title{RFM Table}
\usage{
rfm_table(data = NULL, customer_id = NULL, order_date = NULL,
  revenue = NULL, analysis_date = NULL, recency_bins = 5,
  frequency_bins = 5, monetary_bins = 5, ...)

\method{print}{rfm_table}(x, ...)
}
\arguments{
\item{data}{a data.frame or tibble}

\item{customer_id}{unique id of the customer}

\item{order_date}{date of transaction}

\item{revenue}{revenue from the customer}

\item{analysis_date}{data of analysis}
}
\value{
\code{rfm_table} returns a tibble with the following columns:

\item{customer_id}{unique id of the customer}
\item{date_most_recent}{date of the most recent transaction}
\item{recency_days}{number of days since the most recent transaction}
\item{transaction_count}{total number of transactions of the customer}
\item{revenue}{revenue from the customer}
\item{analysis_date}{date of analysis}
\item{recency_bins}{number of bins for recency}
\item{frequency_bins}{number of bins for frequency}
\item{monetary_bins}{number of bins for monetary}
}
\description{
Table with recency, frequency, monetary and RFM score
}
\examples{
analysis_date <- lubridate::as_datetime('2014-04-01 05:30:00', tz = 'UTC')
rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)
}
