% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmfile.R
\name{RFMfunction}
\alias{RFMfunction}
\title{calculate RFM}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{jsdködv}

\item{weight_recency}{fjdksöjv}

\item{weight_frequency}{jfksöfö}

\item{weight_monetary}{jfaöjöd}
}
\value{

}
\description{
calculate RFM
}
