% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calculateRFM}
\alias{calculateRFM}
\title{calculateRFM}
\usage{
calculateRFM(data, default_r, default_f, default_m)
}
\arguments{
\item{data}{- A data.table containing the transaction rececord details for every customer.}

\item{weight_recency}{- Weight of recency.}

\item{weight_frequency}{- Weight of frequency.}

\item{weight_monetary}{- Weight of monetary.}
}
\value{
Returns a data.data containing the recency, frequency and monetary
score as well as the weighted final score and the group membership.
}
\description{
Calculate the a weighted RFM score: recency, frequency, and monetary for every customers
}
\details{
\code{data} contains the transactional data. The dataset must contain a
column labeled "Customer" that allows unique customer identification
and a column labeled "TransDate", indicating the purchase date.
The column "PurchAmount" specifies the total spending per purchase.
}
\examples{
temp <- RFMfunction(transactions,60,20,20)

}
