% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RFM.R
\name{rfm}
\alias{rfm}
\title{RFM}
\usage{
rfm(data, analysisDate = Sys.Date(), customer = "customer", date = "date",
  revenue = "amount")
}
\arguments{
\item{data}{Dataset for analysis.
Should contain at least three columns : customer name, date of purchase and amount}

\item{analysisDate}{Date used of recency calculations. Defaults to current date.}

\item{customer}{Name of column containing name of customer. Default : customer}

\item{date}{Name of column containing date of transaction. Default : date}

\item{revenue}{Name of column containing the amount of transaction. Default : amount}
}
\description{
RFM (Recency, Frequency, Monetary ) analysis
}
\examples{
rfm(sales)
}
\author{
Syam Murali
}
\keyword{RFM}

