% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm.R
\name{rfm}
\alias{rfm}
\title{rfm_model fucntion}
\usage{
rfm(df, n_freq = c(2:5), n_rec = c(2:5), n_mon = c(2:5), score = TRUE)
}
\arguments{
\item{df}{a data frame loaded with rfm_load or created previously}

\item{n_freq}{an integer between 2 and 5 representing the number of quantile to create for the frequency}

\item{n_rec}{an integer between 2 and 5 representing the number of quantile to create for the recency}

\item{n_mon}{an interget between 2 and 5 representing the number of quantile to create for the monetary score}

\item{score}{a logical default is TRUE which add a concatenated rfm score as a variable in the return object. If FALSE then the score will not be added}
}
\value{
a Data.frame object
}
\description{
Load data in to create model
}

