% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_bar_chart}
\alias{rfm_bar_chart}
\title{RFM Bar Chart}
\usage{
rfm_bar_chart(rfm_table)
}
\arguments{
\item{rfm_table}{an object of class \code{rfm_table}}
}
\value{
Bar Chart
}
\description{
Examine the distribution of monetary scores for the different 
combinations of frequency and recency scores.
}
\examples{
# rfm table
analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# bar chart
rfm_bar_chart(rfm_result)

}
