% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-data.R
\docType{data}
\name{rfm_data_orders}
\alias{rfm_data_orders}
\title{RFM Data Set}
\format{A tibble with 49.6 rows and 3 variables:
\describe{
  \item{order_date}{order date}
  \item{customer_id}{customer id}
  \item{revenue}{transaction amount}
}}
\usage{
rfm_data_orders
}
\description{
A dataset containing transactions of different customers.
}
\keyword{datasets}
