% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_heatmap}
\alias{rfm_heatmap}
\title{RFM Heatmap}
\usage{
rfm_heatmap(data)
}
\arguments{
\item{data}{a data.frame or a tibble}
}
\value{
Heatmap
}
\description{
Heatmap of Mean Monetary Value
}
\examples{
# rfm table
analysis_date <- lubridate::as_datetime('2014-04-01 05:30:00', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# heat map
rfm_heatmap(rfm_result)

}
