% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_histograms}
\alias{rfm_histograms}
\title{RFM Histograms}
\usage{
rfm_histograms(rfm_table)
}
\arguments{
\item{rfm_table}{an object of class \code{rfm_table}}
}
\value{
Histograms
}
\description{
Histograms of recency, frequency and monetary value
}
\examples{
# rfm table
analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# histograms
rfm_histograms(rfm_result)

}
