% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_order_dist}
\alias{rfm_order_dist}
\title{Customers by Orders}
\usage{
rfm_order_dist(rfm_table)
}
\arguments{
\item{rfm_table}{an object of class \code{rfm_table}}
}
\value{
Bar Chart
}
\description{
Visualize the distribution of customers across orders
}
\examples{
# rfm table
analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# bar chart
rfm_order_dist(rfm_result)

}
