% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_rm_plot}
\alias{rfm_rm_plot}
\alias{rfm_fm_plot}
\alias{rfm_rf_plot}
\title{RFM Scatter plot}
\usage{
rfm_rm_plot(
  segments,
  xaxis_label = NULL,
  yaxis_label = NULL,
  plot_title = NULL,
  print_plot = TRUE
)

rfm_fm_plot(
  segments,
  xaxis_label = NULL,
  yaxis_label = NULL,
  plot_title = NULL,
  print_plot = TRUE
)

rfm_rf_plot(
  segments,
  xaxis_label = NULL,
  yaxis_label = NULL,
  plot_title = NULL,
  print_plot = TRUE
)
}
\arguments{
\item{segments}{Output from \code{rfm_segment}.}

\item{xaxis_label}{X axis label.}

\item{yaxis_label}{Y axis label.}

\item{plot_title}{Title of the plot.}

\item{print_plot}{logical; if \code{TRUE}, prints the plot else returns a plot object.}
}
\value{
Scatter plot.
}
\description{
Examine the relationship between recency, frequency
  and monetary values.
}
\section{Deprecated Functions}{

\code{rfm_rm_plot()}, \code{rfm_fm_plot()} and \code{rfm_rf_plot()} have
been deprecated and will be made defunct. These functions have been provided
for compatibility with older versions only, and will be made defunct at the
next release.
Instead use the replacement function \code{rfm_plot_segment_scatter()()}.
}

