% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_rm_plot}
\alias{rfm_rm_plot}
\alias{rfm_fm_plot}
\alias{rfm_rf_plot}
\title{RFM Scatter Plot}
\usage{
rfm_rm_plot(rfm_table)

rfm_fm_plot(rfm_table)

rfm_rf_plot(rfm_table)
}
\arguments{
\item{rfm_table}{an object of class \code{rfm_table}}
}
\value{
Scatter Plot
}
\description{
Recency vs Monetary Value, Frequency vs Monetary Value and 
Recency vs Frequency Plots
}
\examples{
# rfm table
analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# monetary value vs recency
rfm_rm_plot(rfm_result)

# frequency vs monetary value
rfm_fm_plot(rfm_result)

# frequency vs recency
rfm_rf_plot(rfm_result)

}
