% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm-plots.R
\name{rfm_rm_plot}
\alias{rfm_rm_plot}
\alias{rfm_fm_plot}
\alias{rfm_rf_plot}
\title{RFM Scatter Plot}
\usage{
rfm_rm_plot(rfm_table, point_color = "blue", xaxis_title = "Monetary",
  yaxis_title = "Recency", plot_title = "Recency vs Monetary")

rfm_fm_plot(rfm_table, point_color = "blue", xaxis_title = "Monetary",
  yaxis_title = "Frequency", plot_title = "Frequency vs Monetary")

rfm_rf_plot(rfm_table, point_color = "blue", xaxis_title = "Frequency",
  yaxis_title = "Recency", plot_title = "Recency vs Frequency")
}
\arguments{
\item{rfm_table}{an object of class \code{rfm_table}}

\item{point_color}{color of the scatter points}

\item{xaxis_title}{x axis title}

\item{yaxis_title}{y axis title}

\item{plot_title}{title of the plot}
}
\value{
Scatter Plot
}
\description{
Examine the relationship between the above recency, frequency and monetary values
}
\examples{
# rfm table
analysis_date <- lubridate::as_date('2006-12-31', tz = 'UTC')
rfm_result <- rfm_table(rfm_data, customer_id, order_date, revenue, analysis_date)

# monetary value vs recency
rfm_rm_plot(rfm_result)

# frequency vs monetary value
rfm_fm_plot(rfm_result)

# frequency vs recency
rfm_rf_plot(rfm_result)

}
