% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateRFM.R
\name{RFMfunction}
\alias{RFMfunction}
\title{Measuring the value of a customer within the RFM framework}
\usage{
RFMfunction(d, weight_recency = 1, weight_frequency = 1, weight_monetary = 1)
}
\arguments{
\item{arg1}{Customer data set.}

\item{arg2}{A number.}

\item{arg3}{A number.}

\item{arg4}{A number.}
}
\value{
A data set.
}
\description{
Measuring the value of a customer within the RFM framework
}
\details{
\code{data} - Measure the recency, frequency, and monetary value of customer purchases.
            -Set a score to rank customers according to their purchase recency, frequency, and monetary value.
            -Calculate the overall RFM score
}
\examples{
add(1, 1)
add(10, 1)
}
